fix: reuse main annotation logic from code lens in debug provider
This closes #3458 and reuses the logic we had in the run/test code lens
to special case the top level `@main` annotation. We now use the same
logic in both places and it's moved into the `DebugProvider`.
